requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

eventregistry==8.12
ipykernel == 6.15.0
pandas ==1.5.0
openpyxl==3.1.2
azure-functions==1.14.0
eventregistry==8.12
ipykernel==6.15.0
ipython==8.14.0
jupyter_client==8.2.0
jupyter_core==5.3.0
nltk==3.8.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==1.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
PyJWT==2.7.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
snowflake-connector-python==3.0.4
azure-storage-blob==12.16.0

